requirements/local.txt
-r base.txt

Werkzeug[watchdog]==2.1.2 # https://github.com/pallets/werkzeug
ipdb==0.13.9  # https://github.com/gotcha/ipdb
psycopg2-binary==2.9.3  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
mypy==0.961  # https://github.com/python/mypy
django-stubs==1.12.0  # https://github.com/typeddjango/django-stubs
pytest==7.1.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.5  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs==1.7.0  # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx==5.0.2  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2021.3.14 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1  # https://github.com/PyCQA/flake8
flake8-isort==4.1.1  # https://github.com/gforcada/flake8-isort
coverage==6.4.2  # https://github.com/nedbat/coveragepy
black==22.6.0  # https://github.com/psf/black
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pre-commit==2.20.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.5.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.2.0  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==2.0.3  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django
# django ssl-server for development
django-sslserver==0.22
